Handle multiple requests in the socket buffer
More than one request can get into the socket buffer before we read
it. If so, don't treat it as one request and corrupt the data that is
returned to Flex, as in LT-19122.

Just process the first request in the buffer and save the next one to
process later.